app/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
google-cloud-storage==2.9.0
google-cloud-firestore==2.19.0
google-cloud-logging==3.11.3
google-cloud-vision==2.4.2
requests==2.28.2
beautifulsoup4==4.11.1
